LayoutTests/fast/filesystem/workers/file-entry-to-uri-sync.html

<html>
<head>
<script src="../../../resources/js-test-pre.js"></script>
</head>
<body>
<script>
worker = startWorker('../resources/file-entry-to-uri-sync.js');
</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
</html>
