LayoutTests/fast/filesystem/workers/file-entry-to-uri-sync-expected.txt

[Worker] Obtaining URI from FileEntry.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Starting worker: ../resources/file-entry-to-uri-sync.js
PASS [Worker] testFileEntry.toURI() is 'filesystem:file:///temporary/testFileEntry.txt'
PASS successfullyParsed is true

TEST COMPLETE
